test(storage): deflake integration tests
Create fewer buckets during the integration tests. That reduces the quota usage
and (hopefully) deflakes the integration tests.

Refactored the code to make it easier to reuse buckets across tests.
